ENH: changes for windows compatibility of ITKCommandIO and slicerio -- still having a link problem on windows, so ITKCommandIO disabled for now
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=593
git-svn-id: http://svn.slicer.org/Slicer4/trunk@593 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/SlicerIO/utilities.c
@@ -49,7 +49,7 @@ int vaFormatStringLength(const char *format, va_list ap)
     int total;
     char *ptr = (char *)format;
     char *tmp;
-    int minsize = 0;
+    size_t minsize = 0;
 
     /* Start with the length of the format string + 1 for the \0 */
     total = strlen(format) + 1;
